image-search/requirements.txt
gradio
pillow
torch
open-clip-torch
accelerate
qdrant-client
openai
python-dotenv
transformers
qdrant-client